Shortened the file names in the contexts subpackage.
