+++ pydidas/contexts/diff_exp/__init__.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas.
 #
-# Copyright 2023, Helmholtz-Zentrum Hereon
+# Copyright 2024, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -23,7 +23,7 @@ actual imports/exports.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2024, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
@@ -31,32 +31,32 @@ __all__ = []
 
 
 # import __all__ items from modules:
-from .diffraction_experiment import *
-from .diffraction_experiment_context import *
-from .diffraction_experiment_io import *
-from .diffraction_experiment_io_base import *
+from .diff_exp import *
+from .diff_exp_context import *
+from .diff_exp_io import *
+from .diff_exp_io_base import *
 
 # add modules' __all__ items to package's __all__ items and unclutter the
 # namespace by deleting the module references:
-from . import diffraction_experiment
+from . import diff_exp
 
-__all__.extend(diffraction_experiment.__all__)
-del diffraction_experiment
+__all__.extend(diff_exp.__all__)
+del diff_exp
 
-from . import diffraction_experiment_context
+from . import diff_exp_context
 
-__all__.extend(diffraction_experiment_context.__all__)
-del diffraction_experiment_context
+__all__.extend(diff_exp_context.__all__)
+del diff_exp_context
 
-from . import diffraction_experiment_io
+from . import diff_exp_io
 
-__all__.extend(diffraction_experiment_io.__all__)
-del diffraction_experiment_io
+__all__.extend(diff_exp_io.__all__)
+del diff_exp_io
 
-from . import diffraction_experiment_io_base
+from . import diff_exp_io_base
 
-__all__.extend(diffraction_experiment_io_base.__all__)
-del diffraction_experiment_io_base
+__all__.extend(diff_exp_io_base.__all__)
+del diff_exp_io_base
 
 
 # Automatically find and import IO classes to have them registered
@@ -68,10 +68,7 @@ _dir = __os.path.dirname(__file__)
 _io_classes = set(
     item.strip(".py")
     for item in __os.listdir(_dir)
-    if (
-        item.startswith("diffraction_experiment_io")
-        and not item[-7:] in ["base.py", "meta.py"]
-    )
+    if (item.startswith("diff_exp_io") and not item[-7:] in ["base.py", "meta.py"])
 )
 
 for _module in _io_classes:
