+++ pydidas/contexts/diff_exp/diff_exp_context.py
@@ -16,7 +16,8 @@
 # along with Pydidas. If not, see <http://www.gnu.org/licenses/>.
 
 """
-Subpackage with pydidas contexts.
+Module with the DiffractionExperimentContext singleton which is used to manage
+global information about the experiment independant from the individual frames.
 """
 
 __author__ = "Malte Storm"
@@ -24,38 +25,11 @@ __copyright__ = "Copyright 2024, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
-__all__ = []
-
-# import sub-packages:
-from . import diff_exp
-from . import scan
-
-__all__.extend(["diff_exp", "scan"])
+__all__ = ["DiffractionExperimentContext"]
 
 
-# import __all__ items from modules:
+from ...core import SingletonFactory
+from .diff_exp import DiffractionExperiment
 
-# explicitly import the singleton factories from the subpackages
-from .diff_exp import (
-    DiffractionExperiment,
-    DiffractionExperimentContext,
-    DiffractionExperimentIo,
-)
-from .scan import ScanContext, ScanIo, Scan
 
-__all__.extend(
-    [
-        "DiffractionExperimentContext",
-        "DiffractionExperimentIo",
-        "ScanContext",
-        "ScanIo",
-    ]
-)
-
-GLOBAL_CONTEXTS = {
-    "diffraction_experiment_context": DiffractionExperimentContext(),
-    "scan_context": ScanContext(),
-}
-
-# add modules' __all__ items to package's __all__ items and unclutter the
-# namespace by deleting the module references:
+DiffractionExperimentContext = SingletonFactory(DiffractionExperiment)
