+++ pydidas/contexts/diff_exp/diff_exp_io_yaml.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas.
 #
-# Copyright 2023, Helmholtz-Zentrum Hereon
+# Copyright 2024, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -21,7 +21,7 @@ export DiffractionExperimentContext metadata from a YAML file.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2024, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
@@ -36,9 +36,9 @@ import yaml
 
 from ...core import UserConfigError
 from ...core.constants import LAMBDA_IN_A_TO_E, YAML_EXTENSIONS
-from .diffraction_experiment import DiffractionExperiment
-from .diffraction_experiment_context import DiffractionExperimentContext
-from .diffraction_experiment_io_base import DiffractionExperimentIoBase
+from .diff_exp import DiffractionExperiment
+from .diff_exp_context import DiffractionExperimentContext
+from .diff_exp_io_base import DiffractionExperimentIoBase
 
 
 EXP = DiffractionExperimentContext()
