+++ pydidas/workflow/result_io/workflow_result_io_hdf5.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas.
 #
-# Copyright 2023, Helmholtz-Zentrum Hereon
+# Copyright 2023 - 2024, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -21,7 +21,7 @@ Hdf5 file format.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2023 - 2024, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
@@ -37,8 +37,8 @@ import h5py
 
 from ... import VERSION
 from ...contexts import DiffractionExperimentContext, ScanContext
-from ...contexts.diffraction_exp_context import DiffractionExperiment
-from ...contexts.scan_context import Scan
+from ...contexts.diff_exp import DiffractionExperiment
+from ...contexts.scan import Scan
 from ...core import Dataset
 from ...core.constants import HDF5_EXTENSIONS
 from ...core.utils import create_hdf5_dataset, read_and_decode_hdf5_dataset
@@ -380,7 +380,7 @@ class WorkflowResultIoHdf5(WorkflowResultIoBase):
         node_info : dict
             A dictionary with node_label, data_label, plugin_name keys and the
             respective values.
-        scan : pydidas.contexts.scan_context.Scan
+        scan : pydidas.contexts.scan.Scan
             The imported scan configuration.
         diffraction_exp : pydidas.contexts.diffraction_exp.DiffractionExperiment
             The inported diffraction experiment configuration.
