+++ pydidas/workflow/result_io/workflow_result_io_meta.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas.
 #
-# Copyright 2023, Helmholtz-Zentrum Hereon
+# Copyright 2023 - 2024, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -24,7 +24,7 @@ the specified file formats.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2023 - 2024, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
@@ -35,8 +35,8 @@ import os
 from pathlib import Path
 from typing import Union
 
-from ...contexts.diffraction_exp_context import DiffractionExperiment
-from ...contexts.scan_context import Scan
+from ...contexts.diff_exp import DiffractionExperiment
+from ...contexts.scan import Scan
 from ...core import Dataset
 from ...core.io_registry import GenericIoMeta
 from ...core.utils import get_extension
