+++ tests/contexts/diff_exp/test_diff_exp_io_base.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas.
 #
-# Copyright 2023, Helmholtz-Zentrum Hereon
+# Copyright 2023 - 2024, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -18,7 +18,7 @@
 """Unit tests for pydidas modules."""
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2023 - 2024, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
@@ -29,13 +29,11 @@ import shutil
 import tempfile
 import unittest
 
-from pydidas.contexts.diffraction_exp_context import (
+from pydidas.contexts.diff_exp import (
     DiffractionExperiment,
     DiffractionExperimentContext,
 )
-from pydidas.contexts.diffraction_exp_context.diffraction_experiment_io_base import (
-    DiffractionExperimentIoBase,
-)
+from pydidas.contexts.diff_exp.diff_exp_io_base import DiffractionExperimentIoBase
 from pydidas.core import UserConfigError
