+++ tests/contexts/diff_exp/test_diff_exp_io_yaml.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas.
 #
-# Copyright 2023, Helmholtz-Zentrum Hereon
+# Copyright 2023 - 2024, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -18,7 +18,7 @@
 """Unit tests for pydidas modules."""
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2023 - 2024, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
@@ -32,10 +32,8 @@ import unittest
 import yaml
 
 from pydidas.contexts import DiffractionExperimentContext
-from pydidas.contexts.diffraction_exp_context import DiffractionExperiment
-from pydidas.contexts.diffraction_exp_context.diffraction_experiment_io_yaml import (
-    DiffractionExperimentIoYaml,
-)
+from pydidas.contexts.diff_exp import DiffractionExperiment
+from pydidas.contexts.diff_exp.diff_exp_io_yaml import DiffractionExperimentIoYaml
 from pydidas.core import UserConfigError
