+++ tests/contexts/scan/test_scan_io.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas.
 #
-# Copyright 2023, Helmholtz-Zentrum Hereon
+# Copyright 2023 - 2024, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -18,26 +18,24 @@
 """Unit tests for pydidas modules."""
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2023 - 2024, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
 
 
+import os
+import shutil
+import tempfile
 import unittest
 
-from pydidas.contexts.diffraction_exp_context import (
-    DiffractionExperiment,
-    DiffractionExperimentContext,
-    DiffractionExperimentIo,
-    DiffractionExperimentIoBase,
-)
+from pydidas.contexts.scan import Scan, ScanContext, ScanIo, ScanIoBase
 
 
-EXP = DiffractionExperimentContext()
+SCAN = ScanContext()
 
 
-class TestIo(DiffractionExperimentIoBase):
+class TestIo(ScanIoBase):
     extensions = ["test"]
     format_name = "Test"
 
@@ -47,7 +45,6 @@ class TestIo(DiffractionExperimentIoBase):
         cls.exported = False
         cls.export_filename = None
         cls.import_filename = None
-        cls.diffraction_exp = None
 
     @classmethod
     def export_to_file(cls, filename, **kwargs):
@@ -55,51 +52,53 @@ class TestIo(DiffractionExperimentIoBase):
         cls.export_filename = filename
 
     @classmethod
-    def import_from_file(cls, filename, diffraction_exp=None):
+    def import_from_file(cls, filename, scan):
         cls.imported = True
-        cls.diffraction_exp = EXP if diffraction_exp is None else diffraction_exp
+        cls.scan = SCAN if scan is None else scan
         cls.import_filename = filename
 
 
-class TestDiffractionExperimentIo(unittest.TestCase):
+class TestScanIo(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
-        cls._original_registry = DiffractionExperimentIo.registry.copy()
-        DiffractionExperimentIo.clear_registry()
-        DiffractionExperimentIo.register_class(TestIo)
+        cls._original_registry = ScanIo.registry.copy()
+        ScanIo.clear_registry()
+        ScanIo.register_class(TestIo)
 
     @classmethod
     def tearDownClass(cls):
-        DiffractionExperimentIo.registry = cls._original_registry
+        ScanIo.registry = cls._original_registry
 
     def setUp(self):
+        self._tmppath = tempfile.mkdtemp()
         TestIo.reset()
 
     def tearDown(self):
-        EXP.restore_all_defaults(True)
+        shutil.rmtree(self._tmppath)
+        SCAN.restore_all_defaults(True)
 
     def test_export_to_file(self):
-        _fname = "test.test"
-        DiffractionExperimentIo.export_to_file(_fname)
+        _fname = os.path.join(self._tmppath, "test.test")
+        ScanIo.export_to_file(_fname)
         self.assertTrue(TestIo.exported)
         self.assertEqual(TestIo.export_filename, _fname)
 
-    def test_import_from_file__generic(self):
-        _fname = "test.test"
-        DiffractionExperimentIo.import_from_file(_fname)
+    def test_import_from_file__generic_ScanContext(self):
+        _fname = os.path.join(self._tmppath, "test.test")
+        ScanIo.import_from_file(_fname)
         self.assertTrue(TestIo.imported)
         self.assertEqual(TestIo.import_filename, _fname)
-        self.assertEqual(TestIo.diffraction_exp, EXP)
+        self.assertEqual(TestIo.scan, SCAN)
 
-    def test_import_from_file__given_Exp(self):
-        _exp = DiffractionExperiment()
-        _fname = "test.test"
-        DiffractionExperimentIo.import_from_file(_fname, diffraction_exp=_exp)
+    def test_import_from_file__given_Scan(self):
+        _fname = os.path.join(self._tmppath, "test.test")
+        _scan = Scan()
+        ScanIo.import_from_file(_fname, scan=_scan)
         self.assertTrue(TestIo.imported)
         self.assertEqual(TestIo.import_filename, _fname)
-        self.assertEqual(TestIo.diffraction_exp, _exp)
+        self.assertEqual(TestIo.scan, _scan)
 
 
 if __name__ == "__main__":
     unittest.main()
-    DiffractionExperimentIo.clear_registry()
+    ScanIo.clear_registry()
