+++ tests/contexts/scan/test_scan_io_base.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas.
 #
-# Copyright 2023, Helmholtz-Zentrum Hereon
+# Copyright 2023 - 2024, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -18,7 +18,7 @@
 """Unit tests for pydidas modules."""
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2023 - 2024, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
@@ -32,8 +32,8 @@ from numbers import Integral, Real
 
 import numpy as np
 
-from pydidas.contexts.scan_context import Scan, ScanContext
-from pydidas.contexts.scan_context.scan_io_base import ScanIoBase
+from pydidas.contexts.scan import Scan, ScanContext
+from pydidas.contexts.scan.scan_io_base import ScanIoBase
 from pydidas.core import UserConfigError
 from pydidas.core.utils import get_random_string
