+++ tests/plugins/test_base_plugin.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas.
 #
-# Copyright 2023, Helmholtz-Zentrum Hereon
+# Copyright 2023 - 2024, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -18,7 +18,7 @@
 """Unit tests for pydidas modules."""
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2023 - 2024, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
@@ -33,7 +33,7 @@ import unittest
 import numpy as np
 
 from pydidas.contexts import DiffractionExperimentContext
-from pydidas.contexts.diffraction_exp_context import DiffractionExperiment
+from pydidas.contexts.diff_exp import DiffractionExperiment
 from pydidas.core import Parameter, UserConfigError, get_generic_parameter, utils
 from pydidas.core.constants import BASE_PLUGIN
 from pydidas.core.utils import rebin2d
