+++ tests/workflow/test_workflow_results.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas.
 #
-# Copyright 2023, Helmholtz-Zentrum Hereon
+# Copyright 2023 - 2024, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -18,7 +18,7 @@
 """Unit tests for pydidas modules."""
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2023 - 2024, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
@@ -36,8 +36,8 @@ import numpy as np
 from qtpy import QtWidgets
 
 import pydidas
-from pydidas.contexts.diffraction_exp_context import DiffractionExperimentContext
-from pydidas.contexts.scan_context import Scan, ScanContext
+from pydidas.contexts.diff_exp import DiffractionExperimentContext
+from pydidas.contexts.scan import Scan, ScanContext
 from pydidas.core import Dataset, Parameter, UserConfigError, get_generic_parameter
 from pydidas.core.utils import get_random_string
 from pydidas.plugins import PluginCollection
